feat: SHA-256 circuit with 2-phase challenge passes MockProver
